# dashboard/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Membrane API",
    "description": "Open-core proxy and parallel swarm ingestion engine offering L1/L2 semantic caching. Drop-in OpenAI replacement.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://membrane-api.com/v1",
      "description": "Primary API Gateway Tunnel"
    },
    {
      "url": "https://membrane-wh1g.onrender.com/v1",
      "description": "Direct Backend Instance (Render)"
    },
    {
      "url": "http://localhost:8000/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/chat/completions": {
      "post": {
        "summary": "Create chat completion",
        "description": "Proxy endpoint that acts as a drop-in replacement for standard OpenAI completions. Automatically caches prompts and prunes middle conversational history unless context preservation is explicitly requested.",
        "operationId": "createChatCompletion",
        "parameters": [
          {
            "name": "X-Membrane-Preserve-Context",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "false"
            },
            "description": "Set to 'true' to bypass automatic middle conversation pruning and preserve multi-turn history."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string",
                    "default": "membrane-engagement-layer"
                  },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "role",
                        "content"
                      ]
                    }
                  },
                  "use_global_cache": {
                    "type": "boolean",
                    "default": false
                  },
                  "response_format": {
                    "type": "object"
                  },
                  "temperature": {
                    "type": "number",
                    "default": 0.0
                  },
                  "stream": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "required": [
                  "messages"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful completion response with metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  }
}